feat: introduce AbstractResponseHandler
The 'AbstractResponseHandler' is an abstract Handler class
that, working in tandem with the 'Requests' model, allows to handle
responses of a request-response interaction protocol via a callback
mechanism. To handle a specific protocol, it is enough to define
a new handler class that handles the protocol, inherit from
the abstract handler class, and update some class attributes.