fix: return code=1 in default abci responses
In the AbciHandler of skill valory/abstract_abci, for the responses that
contain a return code or result object, we always return error/reject.
The reason is that the client of the default ABCI handler should not
receive a false positive response as the default handler does not implement
an actual ABCI app.