test: fix e2e catch-up test
in some execution of the e2e catch-up test, it might happen that
some behaviours of the delayed agent are not executed, and so
certain strings that are expected to be printed by behaviours might
not be found in the agent stdout.

This commit addresses that by fixing the "CHECK_STRINGS" set of the
price estimation abci e2e catch-up test. In particular,
strings are divided into two categories: "BEHAVIOUR" and "ROUND" strings; all the other
tests can continue to check for all the strings, whereas the e2e
test will only use the "ROUND" strings.

The variable CHECK_STRINGS_LABELLED is a list of pairs, where each pair
is (expected_output_line, string_type). We preferred this over other
data structures (e.g. a dict with type as key) so to make it easier to
add/remove strings and to keep them in the right order according to the test.