fix: handle late messages when period state has changed
Remove `request_id_to_backup_callback` dictionary and `AsyncState.WAITING_UNHANDLED_MESSAGE`, since they cannot be used in order to determine when a message is arriving late. Instead, we use the period state to define whether it has changed, which means that we are handling a message which arrived late.