feat: update keeper selection logic to allow for retrying
If the keepers' subset is not full yet, re-select a keeper as normally.

Otherwise, cycle through the keepers' subset, using the following logic:

A `PENDING` verification status means that we have not received any errors, therefore, all we know is that the tx has not been mined yet due to low pricing. Consequently, we are going to retry with the same keeper in order to replace the transaction. However, if we receive a status other than `PENDING`, we need to cycle through the keepers' subset. Moreover, if the current keeper has reached the allowed number of retries, then we cycle anyway.