fix: message for previously validated tx
The `to_be_validated_tx_hash` property should not be accessed before we initialize the `tx_hashes_history`. This could be the case if we time out during the first attempt to finalize. Therefore, we will not have any history at all, apart from the late arriving hash which has been synchronized, and we have just found out that has been verified.