fix: increase observation interval for tests that reset tm
When resetting with Tendermint, the `observation_interval/2` is used in order to wait when entering and exiting the reset round. However, in #904 we updated the observation interval for all the tests to 3. As a result, we only wait for 1.5sec when entering the reset round before the reset begins. This may result to the **minority** of the agents to reset and when they enter back they cannot continue. Therefore, this commit increases the observation interval for the tests that also reset Tendermint nodes in order to allow the majority of the agents to reset properly.

However, this is limiting for the developer because it cannot use a very small value for the observation interval.