refactor: use the round checks to define the happy path
We now need to define a happy path for our e2e tests, using the `RoundChecks` dataclass. This enables us to specify the name of the successful event which previously could only be `DONE` for the test to succeed. This will also allow us to re-enable the APY e2e tests.