fix: replace Popen.wait with Popen.communicate in abci connection (#1073)
This fix is about a known problem with `subprocesses.Popen` when `stdout=PIPE` or `stderr=PIPE` is used:
https://docs.python.org/3/library/subprocess.html#subprocess.Popen.wait

> Note: This will deadlock when using `stdout=PIPE` or `stderr=PIPE` and the child process generates enough output to a pipe such that it blocks waiting for the OS pipe buffer to accept more data. Use [`Popen.communicate()`](https://docs.python.org/3/library/subprocess.html#subprocess.Popen.communicate) when using pipes to avoid that.

And indeed ABCI connection uses `Popen.wait` to wait for the subprocess termination.

On the other hand, the Python documentation of `subprocess` gives another warning related to the use of `Popen.communicate()` (https://docs.python.org/3/library/subprocess.html#subprocess.Popen.communicate):

> Note: The data read is buffered in memory, so do not use this method if the data size is large or unlimited.

We should be OK for our use case; still, good to keep it in mind.