test: use aea.helpers.io.open_file for dummy service files
For correct cross-platform compatibility of the test_publish_and_fetch_service_ipfs
test function, we need to make sure the files of the dummy AEA service are
saved with UNIX newlines. Before this commit, this was not the case because
the Python built-in 'open' function was used to create the new files; however,
by default, the newline keyword-argument is set to None, which means the
actual newline character used to split lines depends from the underlying
operating system.

This commit changes this and replaces the usage of 'open' with aea.helpers.io.open_file
helper function. In this way, the test should be compatible with Windows platforms.