test: replace check string
The registration e2e tests sometimes fail. The issue is because of the way that we check if the test succeeded, and not because the test does not really succeed. We are checking the logs of the agent. However, if the execution reaches a degenerate round, then it starts looping and logging very fast the same exception. This results in the test not detecting the expected logs - probably because of a buffer filling - even though they have been logged, and falsely reporting the e2e test to have failed.

Therefore, we replace the check string with another one that happens a bit before reaching the degenerate round, but still ensures that the test has succeeded.

However, this solution does not guarantee that the e2e test will not falsely report a failure again, but decreases the chances.

Two more appropriate solutions would be to:
 1. Decrease the interval of checking the logs.
 2. Raise and stop the execution when a degenerate round is reached instead of looping.