chore: increase validate timeout by an order of magnitude
We do this because the timeout was not enough in most of the cases. For example take a look below:
```
[2022-09-12 18:10:08,049] [INFO] [agent] Verified result: True
[2022-09-12 18:10:08,050] [INFO] [agent] Finalized with transaction hash: 0xb73dca111e1f33845235d21478a34c20c6166ef713a6c701d1306e6604ff0e3f
[2022-09-12 18:10:08,409] [INFO] [agent] arrived block with timestamp: 2022-09-12 18:10:05.043275
[2022-09-12 18:10:08,409] [INFO] [agent] current AbciApp time: 2022-09-12 18:09:59.054532
[2022-09-12 18:10:08,409] [WARNING] [agent] expired deadline 2022-09-12 18:10:03.962161 with event Event.VALIDATE_TIMEOUT at AbciApp time 2022-09-12 18:10:05.043275
[2022-09-12 18:10:08,410] [INFO] [agent] current AbciApp time after expired deadline: 2022-09-12 18:10:05.043275
[2022-09-12 18:10:08,410] [INFO] [agent] 'validate_transaction' round is done with event: Event.VALIDATE_TIMEOUT
```

Increasing the timeout by an order of magnitude will give the rounds plenty of time to look up for the tx but at the same time keep the tx settlement transitions fast enough in case that the validation times out in edge cases.