fix: improve consistency of check_abci_app_specs
Now the command check-app-specs also accepts a value to --packages-dir, which
is not necessarily a subdirectory of the current working directory. Without the
change, the command would have failed as the call to
`importlib.import_module("packages...")` could not find the ${packages_dir}
folder in the Python system path.

Another requirement is that the packages directory is named 'packages'. Again,
otherwise, imports of modules whose root module is 'packages' would have
failed. This limitation could be overcome by artificially populating
`sys.modules` with key "packages" and value "{packages_dir}/__init__.py".