fix: get the validator's address as all capital hex
We do not need to decode, but to retrieve the bytes as a hex string representation.

We need to convert to upper because this is how the `validator_to_agent` contains the validators' addresses in its keys. That is because the dictionary is set during registration startup (via the `setup_slashing` method) using the values returned by `/params`. The `priv_validator_key.json` contains it as capital hex, and this is where the `/params` of the tm com server gets it from.