chore: remove `continue-on-error` so the default applies
The default value for the `continue-on-error` flag is `false`. Because the flag was set to `true` for the unit tests, a green status was being reported despite them failing due to an infinite loop introduced in the dev mode in #2221. As a result, the failure went unnoticed for 5 months, which eventually led to additional bugs in the CLI, since its tests appeared to pass.